[Snap Conversion API] Remove the V2 implementation
